# protoc-gen-go-gin/example/api/product/app/v1/v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "product/app/v1/v1.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BlogService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/author/{author_id}/articles": {
      "get": {
        "operationId": "BlogService_GetArticles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArticlesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "author_id",
            "description": "@inject_tag: bind:\"uri,name=author_id\" validate:\"required,gt=0\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "description": "@inject_tag: bind:\"query,name=title\" validate:\"max_len=200\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "@inject_tag: bind:\"query,name=page\" validate:\"gt=0\" default:\"1\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_size",
            "description": "@inject_tag: bind:\"query,name=page_size\" validate:\"gt=0,lte=100\" default:\"20\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "BlogService_CreateArticle",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateArticleResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "author_id",
            "description": "@inject_tag: bind:\"uri,name=author_id\" validate:\"required,gt=0\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "authorization",
            "description": "@inject_tag: bind:\"header,name=Authorization\" validate:\"required\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "title",
            "description": "@inject_tag: bind:\"form,name=title\" validate:\"required,min_len=1,max_len=400\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "content",
            "description": "@inject_tag: bind:\"form,name=content\" validate:\"required,min_len=1,max_len=100000\"",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/images/{key}": {
      "get": {
        "operationId": "BlogService_GetImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetImageResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "description": "@inject_tag: bind:\"uri,name=key\" validate:\"required,lt=128,pattern=#abc123\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "BlogService_UploadImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UploadImageResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "description": "@inject_tag: bind:\"uri,name=key\" validate:\"required,lt=128,pattern=#abc123\"",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "description": "@inject_tag: bind:\"header,name=Authorization\" validate:\"required\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "image",
            "description": "@inject_tag: bind:\"file,name=image\" validate:\"gt=0,lte=102400\"",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          },
          {
            "name": "filename",
            "description": "@inject_tag: bind:\"form,name=filename\"",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Article": {
      "type": "object",
      "properties": {
        "article_id": {
          "type": "integer",
          "format": "int32"
        },
        "author_id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "v1CreateArticleResp": {
      "type": "object",
      "properties": {
        "article_id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "v1GetArticlesResp": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64"
        },
        "articles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Article"
          }
        }
      }
    },
    "v1GetImageResp": {
      "type": "object",
      "properties": {
        "file": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "v1UploadImageResp": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        }
      }
    }
  }
}